Moved code into a new module called `sknn.nn` which will contain everything common to MLP and AE.
